Rename tools/submit.py to validate.py
