+++ src/pkg/caendr/caendr/models/job_pipeline/nemascan_pipeline.py
@@ -8,7 +8,7 @@ from caendr.models.run             import NemascanRunner
 
 # Services
 from caendr.models.datastore       import Species
-from caendr.services.tools.submit  import validate_file, validate_num, validate_strain
+from caendr.services.validate      import validate_file, validate_num, validate_strain
 from caendr.services.cloud.storage import upload_blob_from_file
 from caendr.utils.env              import get_env_var
 from caendr.utils.file             import get_file_hash
